requirements.txt
web3
streamlit
pandas
python-dotenv
tqdm
numpy
pyyaml